# DotPager: Replace useResizeObserver by native resize event