@jsnajdr 2021-11-17
Is there anything wrong with the `useResizeObserver` hook? I don't understand what was the cause of the performance issues and why manual `resize` listener solves them.