@arthur791004 2021-11-18
I think the reason is that `useResizeObserver` uses the `iframe` to detect the size of the page changed or not. However, the height of the page changed as the height of each slide might be different. Therefore, the `resize` event triggers all the time when the height of the `DotPager` changes as we switch to the next slide.